+++ test/pipeline.test.ts
@@ -13,6 +13,7 @@ const pluginDefaults: PluginConfigValues = {
   showSlideNumber: true,
   showProgressBar: true,
   verticalSlides: false,
+  autoRefreshWhilePresenting: true,
   showSampleCommand: true
 }